Separate CDC stuff from generic Kafka Connnect
